unify resource matchers
