+++ src/main/scala/mesosphere/mesos/PersistentVolumeMatcher.scala
@@ -16,16 +16,19 @@ object PersistentVolumeMatcher {
         resource.getDisk.getPersistence.getId -> resource
     }.toMap
 
-    def resourcesForTask(task: Task.Reserved): Option[Iterable[Mesos.Resource]] = {
-      if (task.reservation.volumeIds.map(_.idString).forall(availableVolumes.contains))
-        Some(task.reservation.volumeIds.flatMap(id => availableVolumes.get(id.idString)))
+    def resourcesForTask(volumeIds: Iterable[String]): Option[Iterable[Mesos.Resource]] = {
+      if (volumeIds.forall(availableVolumes.contains))
+        Some(volumeIds.flatMap(id => availableVolumes.get(id)))
       else
         None
     }
 
-    waitingTasks.toStream
-      .flatMap { task => resourcesForTask(task).flatMap(rs => Some(VolumeMatch(task, rs))) }
-      .headOption
+    waitingTasks.toStream.
+      flatMap { task =>
+        resourcesForTask(task.reservation.volumeIds.map(_.idString)).
+          flatMap(rs => Some(VolumeMatch(task, rs)))
+      }.
+      headOption
   }
 
   case class VolumeMatch(task: Task.Reserved, persistentVolumeResources: Iterable[Mesos.Resource])
