+++ src/main/scala/mesosphere/mesos/ResourceMatcher.scala
@@ -1,26 +1,59 @@
 package mesosphere.mesos
 
 import mesosphere.marathon.core.launcher.impl.TaskLabels
-import mesosphere.marathon.state.{ PersistentVolume, ResourceRole, DiskType, DiskSource }
-import mesosphere.marathon.tasks.{ PortsMatcher, ResourceUtil }
-import mesosphere.mesos.protos.Resource
+import mesosphere.marathon.state.{ PersistentVolume, ResourceRole, DiskSource }
+import mesosphere.marathon.tasks.ResourceUtil
 import org.apache.mesos.Protos
-import org.apache.mesos.Protos.Resource.DiskInfo.Source
 import org.apache.mesos.Protos.Offer
 import org.slf4j.LoggerFactory
 
 import scala.annotation.tailrec
 import scala.collection.JavaConverters._
 import scala.collection.immutable.Seq
-import scala.collection.immutable
 import scala.collection.mutable
+import scala.collection
+
+trait ResourceMatcher extends ((String, Iterable[Protos.Resource]) => Iterable[MatchResult]) {
+  val resourceName: ResourceMatcher.Role
+  def apply(offerId: String, resources: Iterable[Protos.Resource]): Iterable[MatchResult]
+}
 
 object ResourceMatcher {
-  import ResourceUtil.RichResource
   type Role = String
   type OfferMatcher[T] = (Offer, ResourceSelector) => T
 
   private[this] val log = LoggerFactory.getLogger(getClass)
+
+  /**
+    * Checks whether the given offer contains enough resources to launch a task of the given run spec
+    * or to make a reservation for a task.
+    *
+    * If a task uses local volumes, this method is typically called twice for every launch. Once
+    * for the reservation on UNRESERVED resources and once for every (re-)launch on RESERVED resources.
+    *
+    * If matching on RESERVED resources as specified by the ResourceSelector, resources for volumes
+    * have to be matched separately (e.g. by the [[PersistentVolumeMatcher]]). If matching on UNRESERVED
+    * resources, the disk resources for the local volumes are included since they must become part of
+    * the reservation.
+    */
+  def matchResources(offer: Offer, matchers: Iterable[ResourceMatcher], selector: ResourceSelector): Option[ResourceMatch] = {
+    require(matchers.map(_.resourceName)(collection.breakOut).size == matchers.size, "overlapping matchers not allowed")
+
+    val groupedResources: Map[Role, mutable.Buffer[Protos.Resource]] = offer.getResourcesList.asScala.groupBy(_.getName)
+    val offerId = offer.getId.getValue
+
+    val matchResults = matchers.
+      flatMap{ matcher => matcher.apply(offerId, groupedResources.getOrElse(matcher.resourceName, Nil)) }.
+      filter(_.wantsResources)
+
+    logUnsatisfiedResources(offer, selector, matchResults)
+
+    if (matchResults.forall(_.matches)) {
+      Some(ResourceMatch(matchResults))
+    } else {
+      None
+    }
+  }
 
   /**
     * A successful match result of the [[ResourceMatcher]].matchResources method.
@@ -130,268 +163,6 @@ object ResourceMatcher {
     }
   }
 
-  case class ResourceRequests(
-      cpus: Double, mem: Double, gpus: Int, disk: Double, persistentVolumes: Seq[PersistentVolume]) {
-    lazy val diskForPersistentVolumes = persistentVolumes.map { _.persistent.size }.sum
-
-  }
-
-  /**
-    * Checks whether the given offer contains enough resources to launch a task of the given run spec
-    * or to make a reservation for a task.
-    *
-    * If a task uses local volumes, this method is typically called twice for every launch. Once
-    * for the reservation on UNRESERVED resources and once for every (re-)launch on RESERVED resources.
-    *
-    * If matching on RESERVED resources as specified by the ResourceSelector, resources for volumes
-    * have to be matched separately (e.g. by the [[PersistentVolumeMatcher]]). If matching on UNRESERVED
-    * resources, the disk resources for the local volumes are included since they must become part of
-    * the reservation.
-    */
-  def matchResources(offer: Offer, req: ResourceRequests, selector: ResourceSelector,
-    portsMatcher: PortsMatcher): Option[ResourceMatch] = {
-
-    val groupedResources: Map[Role, mutable.Buffer[Protos.Resource]] = offer.getResourcesList.asScala.groupBy(_.getName)
-
-    val scalarResourceMatch = matchScalarResource(groupedResources, selector) _
-    val diskResourceMatch = matchDiskResource(groupedResources, selector) _
-
-    // Local volumes only need to be matched if we are making a reservation for resident tasks --
-    // that means if the resources that are matched are still unreserved.
-    def needToReserveDisk = selector.needToReserve && req.diskForPersistentVolumes > 0
-
-    val diskMatch = if (needToReserveDisk)
-      diskResourceMatch(
-        req.disk,
-        req.persistentVolumes,
-        ScalarMatchResult.Scope.IncludingLocalVolumes)
-    else
-      diskResourceMatch(req.disk, Nil, ScalarMatchResult.Scope.ExcludingLocalVolumes)
-
-    val matchResults = (
-      Iterable(
-        portsMatcher(offer),
-        scalarResourceMatch(Resource.CPUS, req.cpus, ScalarMatchResult.Scope.NoneDisk),
-        scalarResourceMatch(Resource.MEM, req.mem, ScalarMatchResult.Scope.NoneDisk),
-        scalarResourceMatch(Resource.GPUS, req.gpus.toDouble, ScalarMatchResult.Scope.NoneDisk)) ++
-        diskMatch
-    ).filter(_.wantsResources)
-
-    logUnsatisfiedResources(offer, selector, matchResults)
-
-    if (matchResults.forall(_.matches)) {
-      Some(ResourceMatch(matchResults))
-    } else {
-      None
-    }
-  }
-
-  private[mesos] case class SourceResources(source: Option[Source], resources: List[Protos.Resource]) {
-    lazy val size = resources.foldLeft(0.0)(_ + _.getScalar.getValue)
-  }
-  private[mesos] object SourceResources extends ((Option[Source], List[Protos.Resource]) => SourceResources) {
-    def listFromResources(l: List[Protos.Resource]): List[SourceResources] = {
-      l.groupBy(_.getDiskSourceOption).map(SourceResources.tupled).toList
-    }
-  }
-
-  /*
-   Prioritize resources to make the most sensible allocation.
-   - If requesting full disk, allocate the smallest disk volume that meets constraints
-   - If requesting root, just return it because there can only be one.
-   - If requesting path disk, allocate the largest volume possible to spread allocation evenly.
-   This may not be ideal if you'd prefer to leave room for larger allocations instead.
-
-   TODO - test. Also, parameterize?
-   */
-  private[this] def prioritizeDiskResources(
-    diskType: DiskType,
-    resources: List[SourceResources]): List[SourceResources] = {
-    diskType match {
-      case DiskType.Root =>
-        resources
-      case DiskType.Path =>
-        resources.sortBy(_.size)(implicitly[Ordering[Double]].reverse)
-      case DiskType.Mount =>
-        resources.sortBy(_.size)
-    }
-  }
-
-  // format: OFF
-  @tailrec
-  private[this] def findDiskGroupMatches(
-    requiredValue: Double,
-    resourcesRemaining: List[SourceResources],
-    allSourceResources: List[SourceResources],
-    matcher: Protos.Resource => Boolean):
-      Option[(Option[Source], List[GeneralScalarMatch.Consumption], List[SourceResources])] =
-    // format: ON
-    resourcesRemaining match {
-      case Nil =>
-        None
-      case next :: rest =>
-        consumeResources(requiredValue, next.resources, matcher = matcher) match {
-          case Left(_) =>
-            findDiskGroupMatches(requiredValue, rest, allSourceResources, matcher)
-          case Right((resourcesConsumed, remainingAfterConsumption)) =>
-            val sourceResourcesAfterConsumption = if (remainingAfterConsumption.isEmpty)
-              None
-            else
-              Some(next.copy(resources = remainingAfterConsumption))
-
-            Some((
-              next.source,
-              resourcesConsumed,
-              (sourceResourcesAfterConsumption ++ (allSourceResources.filterNot(_ == next))).toList))
-        }
-    }
-
-  /**
-    * Match volumes against disk resources and return results which keep disk sources and persistentVolumes associated.
-    *
-    * TODO - handle matches for a single volume across multiple resource offers for the same disk
-    */
-  private[this] def matchDiskResource(
-    groupedResources: Map[Role, mutable.Buffer[Protos.Resource]], selector: ResourceSelector)(
-    scratchDisk: Double,
-    volumes: Iterable[PersistentVolume],
-    scope: ScalarMatchResult.Scope = ScalarMatchResult.Scope.NoneDisk): Seq[ScalarMatchResult] = {
-
-    @tailrec
-    def findMatches(
-      diskType: DiskType,
-      pendingAllocations: List[Either[Double, PersistentVolume]],
-      resourcesRemaining: List[SourceResources],
-      resourcesConsumed: List[DiskResourceMatch.Consumption] = Nil): Either[DiskResourceNoMatch, DiskResourceMatch] = {
-      val orderedResources = prioritizeDiskResources(diskType, resourcesRemaining)
-
-      pendingAllocations match {
-        case Nil =>
-          Right(DiskResourceMatch(diskType, resourcesConsumed, scope))
-        case nextAllocation :: restAllocations =>
-          val (matcher, nextAllocationSize) = nextAllocation match {
-            case Left(size) => ({ _: Protos.Resource => true }, size)
-            case Right(v) => (
-              VolumeConstraints.meetsAllConstraints(_: Protos.Resource, v.persistent.constraints),
-              v.persistent.size.toDouble
-            )
-          }
-
-          findDiskGroupMatches(nextAllocationSize, orderedResources, orderedResources, matcher) match {
-            case None =>
-              Left(
-                DiskResourceNoMatch(resourcesConsumed, resourcesRemaining.flatMap(_.resources), nextAllocation, scope))
-            case Some((source, generalConsumptions, decrementedResources)) =>
-              val consumptions = generalConsumptions.map { c =>
-                DiskResourceMatch.Consumption(c, source, nextAllocation.right.toOption)
-              }
-
-              findMatches(
-                diskType,
-                restAllocations,
-                decrementedResources,
-                consumptions ++ resourcesConsumed)
-          }
-      }
-    }
-
-    /*
-      * The implementation for finding mount matches differs from disk matches because:
-      * - A mount volume cannot be partially allocated. The resource allocation request must be sized up to match the
-      *   actual resource size
-      * - The mount volume can't be split amongst reserved / non-reserved.
-      * - The mount volume has an extra maxSize concern
-      *
-      * If this method can be generalized to worth with the above code, then so be it.
-      */
-    @tailrec def findMountMatches(
-      pendingAllocations: List[PersistentVolume],
-      resources: List[Protos.Resource],
-      resourcesConsumed: List[DiskResourceMatch.Consumption] = Nil): Either[DiskResourceNoMatch, DiskResourceMatch] = {
-      pendingAllocations match {
-        case Nil =>
-          Right(DiskResourceMatch(DiskType.Mount, resourcesConsumed, scope))
-        case nextAllocation :: restAllocations =>
-          resources.find { resource =>
-            val resourceSize = resource.getScalar.getValue
-            VolumeConstraints.meetsAllConstraints(resource, nextAllocation.persistent.constraints) &&
-              (resourceSize >= nextAllocation.persistent.size) &&
-              (resourceSize <= nextAllocation.persistent.maxSize.getOrElse(Long.MaxValue))
-          } match {
-            case Some(matchedResource) =>
-              val consumedAmount = matchedResource.getScalar.getValue
-              val grownVolume =
-                nextAllocation.copy(
-                  persistent = nextAllocation.persistent.copy(
-                    size = consumedAmount.toLong))
-              val consumption =
-                DiskResourceMatch.Consumption(
-                  consumedAmount,
-                  role = matchedResource.getRole,
-                  reservation = if (matchedResource.hasReservation) Option(matchedResource.getReservation) else None,
-                  source = DiskSource.fromMesos(matchedResource.getDiskSourceOption),
-                  Some(grownVolume))
-
-              findMountMatches(
-                restAllocations,
-                resources.filterNot(_ == matchedResource),
-                consumption :: resourcesConsumed)
-            case None =>
-              Left(DiskResourceNoMatch(resourcesConsumed, resources, Right(nextAllocation), scope))
-          }
-      }
-    }
-
-    val diskResources = groupedResources.getOrElse(Resource.DISK, Iterable.empty)
-
-    val resourcesByType: immutable.Map[DiskType, Iterable[Protos.Resource]] = diskResources.groupBy { r =>
-      DiskSource.fromMesos(r.getDiskSourceOption).diskType
-    }.withDefault(_ => Nil)
-
-    val scratchDiskRequest = if (scratchDisk > 0.0) Some(Left(scratchDisk)) else None
-    val requestedResourcesByType: immutable.Map[DiskType, Iterable[Either[Double, PersistentVolume]]] =
-      (scratchDiskRequest ++ volumes.map(Right(_)).toList).groupBy {
-        case Left(_) => DiskType.Root
-        case Right(p) => p.persistent.`type`
-      }
-
-    requestedResourcesByType.keys.map { diskType =>
-      val withBiggestRequestsFirst =
-        requestedResourcesByType(diskType).
-          toList.
-          sortBy({ r => r.right.map(_.persistent.size.toDouble).merge })(implicitly[Ordering[Double]].reverse)
-
-      val resources = resourcesByType(diskType).filter(selector(_)).toList
-
-      if (diskType == DiskType.Mount) {
-        findMountMatches(
-          withBiggestRequestsFirst.flatMap(_.right.toOption),
-          resources)
-      } else
-        findMatches(
-          diskType,
-          withBiggestRequestsFirst,
-          SourceResources.listFromResources(resources))
-    }.toList.map(_.merge)
-  }
-
-  private[this] def matchScalarResource(
-    groupedResources: Map[Role, mutable.Buffer[Protos.Resource]], selector: ResourceSelector)(
-    name: String, requiredValue: Double,
-    scope: ScalarMatchResult.Scope = ScalarMatchResult.Scope.NoneDisk): ScalarMatchResult = {
-
-    require(scope == ScalarMatchResult.Scope.NoneDisk || name == Resource.DISK)
-
-    val resourcesForName = groupedResources.getOrElse(name, Iterable.empty)
-    val matchingScalarResources = resourcesForName.filter(selector(_))
-    consumeResources(requiredValue, matchingScalarResources.toList) match {
-      case Left(valueLeft) =>
-        NoMatch(name, requiredValue, requiredValue - valueLeft, scope = scope)
-      case Right((resourcesConsumed, remaining)) =>
-        GeneralScalarMatch(name, requiredValue, resourcesConsumed, scope = scope)
-    }
-  }
-
   /**
     * Given a list of resources, allocates the specified size.
     *
@@ -402,7 +173,7 @@ object ResourceMatcher {
     *     allocation.
     */
   @tailrec
-  private[this] def consumeResources(
+  private[mesos] def consumeResources(
     valueLeft: Double,
     resourcesLeft: List[Protos.Resource],
     resourcesNotConsumed: List[Protos.Resource] = Nil,
